+++ app/src/main/java/com/example/stellarvision/MainActivity.kt
@@ -1,21 +1,12 @@
 package com.example.stellarvision
 
-import android.hardware.Sensor
 import android.hardware.SensorManager
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.Navigation
-import com.example.stellarvision.Navigation.Navigation
-import com.example.stellarvision.ui.theme.StellarVisionTheme
+import com.example.stellarvision.navigation.Navigation
 
 lateinit var sensorManager: SensorManager
