+++ app/src/main/java/com/example/stellarvision/screens/homepage.kt
@@ -1,5 +1,4 @@
-package com.example.stellarvision.Screens
-
+package com.example.stellarvision.screens
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -13,28 +12,31 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import com.example.stellarvision.common.ActivityTopRow
 import com.example.stellarvision.common.BottomBar
-import com.example.stellarvision.common.NotificationCard
+import com.example.stellarvision.common.HomeTopTabs
+import com.example.stellarvision.common.PostCard
 import com.example.stellarvision.common.iconsNavBar
 
 @Composable
-fun Actividad(controller: NavController)
-{
+fun Homepage(controller: NavController) {
     BottomBar(controller, iconsNavBar) { padding ->
         var selected by rememberSaveable { mutableIntStateOf(1) }
         Column(
             modifier = Modifier.padding(padding).fillMaxSize()
         ) {
-            ActivityTopRow(
+            HomeTopTabs(
                 selectedIndex = selected,
                 modifier = Modifier,
                 onSelect = {selected=it}
             )
-            NotificationCard(
-                userName = "Felipe",
-                groupText = "Hace 1 dia",
-                body = "Le dio like a tu post"
+            PostCard(
+                userName = "Helena",
+                groupText = "Hace 5 minutos",
+                body = "Hoy en la noche tome una foto de la constelacion Orion. Me encanta como se ven las constelaciones con esta aplicacion!",
+                likes= 8,
+                comments = 1,
+                previewUser = "Daniel",
+                previewText = "Se ve bastante bien, desde que lugar la tomaste?"
             )
         }
     }
@@ -42,6 +44,6 @@ fun Actividad(controller: NavController)
 
 @Preview(showBackground = true)
 @Composable
-fun ActividadPreview() {
-    Actividad(controller = rememberNavController())
+fun HomepagePreview() {
+    Homepage(controller = rememberNavController())
 }
